Java, typeres: move method typeresolution methods into a helper class
